adapt for long int by changing to ulong

+++ PizzaSlices/PizzaSlices/SliceFinder.cs
@@ -8,14 +8,14 @@ namespace PizzaSlices
     public class SliceFinder
     {
         private readonly int[] _source;
-        private readonly int _totalPieces;
+        private readonly ulong _totalPieces;
 
         private int[] _previousChosenCombination = new List<int>().ToArray();
         private int[] _previousChosenIndexes = new List<int>().ToArray();
-        private int _previousChosenSum = 0;
+        private ulong _previousChosenSum = 0;
         int _index = 0;
 
-        public SliceFinder(int[] source, int totalPieces)
+        public SliceFinder(int[] source, ulong totalPieces)
         {
             _source = source;
             _totalPieces = totalPieces;
@@ -28,12 +28,12 @@ namespace PizzaSlices
             for (; combinationFor > 0; combinationFor--)
             {
                 var combinationOfSmallestNumbers = _source.Take(combinationFor);
-                var smallSum = combinationOfSmallestNumbers.Aggregate(0, (a, c) => a + c);
+                var smallSum = combinationOfSmallestNumbers.Aggregate(0UL, (a, c) => a + (ulong)c);
                 if (smallSum > _totalPieces) continue;
 
                 foreach (var combination in CombinationGenerator.Generate(_source, combinationFor))
                 {
-                    var sum = combination.Item1.Aggregate(0, (a, c) => a + c);
+                    var sum = combination.Item1.Aggregate(0UL, (a, c) => a + (ulong)c);
                     UpdateState(combination, sum, combinationFor);
 
                     if (CanStop())
@@ -58,7 +58,7 @@ namespace PizzaSlices
             return _previousChosenSum == _totalPieces;
         }
 
-        private void UpdateState(Tuple<int[], int[]> combination, int sum, int index)
+        private void UpdateState(Tuple<int[], int[]> combination, ulong sum, int index)
         {
             if (sum == _totalPieces || sum < _totalPieces && sum > _previousChosenSum)
             {
